+++ packages/dash-core-components/tests/unit/Input.test.js
@@ -46,6 +46,8 @@ describe('Props can be set properly', () => {
             component_name: '',
             prop_name: '',
         },
+        persisted_props: ['value'],
+        persistence_type: 'local',
     };
     const input = mount(<Input {...defaultProps} />);