+++ packages/dash-core-components/tests/unit/RadioItems.test.js
@@ -29,6 +29,8 @@ describe('Props can be set properly', () => {
             component_name: '',
             prop_name: '',
         },
+        persisted_props: ['value'],
+        persistence_type: 'local',
     };
 
     const radio = mount(<RadioItems {...testProps} />);